backend/requirements.txt
#Core Django requirements
Django==5.2.9
asgiref==3.11.0
djangorestframework==3.16.1
python-dotenv==1.2.1
sqlparse==0.5.5

#Database requirements
psycopg2-binary>=2.9.9
dj-database-url>=1.2.0

#Channels requirements
daphne>=4.1.0
channels>=4.1.0

pillow>=10.4.0
twilio>=8.14.0